Refactor to define a place for shared code between training and serving images
