+++ truss/constants.py
@@ -15,7 +15,11 @@ CODE_DIR = pathlib.Path(BASE_DIR, "truss")
 
 TEMPLATES_DIR = pathlib.Path(CODE_DIR, "templates")
 SERVER_CODE_DIR = TEMPLATES_DIR / "server"
-TRAINING_CODE_DIR = TEMPLATES_DIR / "training"
+TRAINING_JOB_WRAPPER_CODE_DIR = TEMPLATES_DIR / "training"
+SHARED_SERVING_AND_TRAINING_CODE_DIR_NAME = "shared"
+SHARED_SERVING_AND_TRAINING_CODE_DIR = (
+    TEMPLATES_DIR / SHARED_SERVING_AND_TRAINING_CODE_DIR_NAME
+)
 CONTROL_SERVER_CODE_DIR = TEMPLATES_DIR / "control"
