+++ truss/templates/shared/README.md
@@ -1,0 +1,3 @@
+# Shared code between training and serving images
+
+Code in this directory is common to both training and serving and is copied into them.
