+++ truss/templates/shared/secrets_resolver.py
